refactor: relocate COC banner to ChatInput and restore country name in messages
Move the country-of-concern banner from the chat transcript area into a card
above the chat input so it stays anchored near the model selector. Re-introduce
the country name in the short and long warning messages, and force COC warning
icons to yellow across the model selector, chat input, and prompt editor.

Removes the countriesOfConcern unit test suite alongside the helper changes.
